Fix Json escaping (#139)
