+++ .gitignore
@@ -10,6 +10,7 @@ target/
 lib_managed/
 src_managed/
 project/boot/
+project/target/
 project/plugins/project/
 
 # Scala-IDE specific
